Final changes pilot report, menu and maintenance use cases

+++ gui/java gui playground/src/Object/FoodMenu.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public class FoodMenu {
     public List<FoodMenuItem> foodMenu;
-    public int menuPrice;
+    public float menuPrice;
     public int menuID;
     public String description;
     public String name;
